scrapynews: Move shell command execution into python
This will make scrapynews' execution flow identical to roadrunner and
webstemmer and will rely on the same FOLDER_NAMES variable for the
input folders we want to analyse.
Since we are no longer appending to python path, a symlink needed to
be added for constants.py so that it may be used in scrapynews.

Signed-off-by: baloncek2662 <martic.gasper@gmail.com>